requirements.txt
cassandra-driver==3.14.0